More work on GUI and Class Definitions
Need to finish the Product definition and add in more checks in the UI.  For example, need to reset itme list and add a warning if it's already been set.  Also, move the company definitions to the Company class to clean up the main function code and add in more comments.